+++ packages/sdk/CHANGELOG.md
@@ -9,7 +9,11 @@ Not every commit needs to result in a change to this file (e.g. docs and chore c
 commit that affects the code in a way that consumers might care about should include edits to the
 'Unreleased' section though. Breaking changes should be prefixed with `**BREAKING:**`.
 
-## [Unreleased](https://github.com/airtable/blocks/compare/@airtable/blocks@1.10.2...HEAD)
+## [Unreleased](https://github.com/airtable/blocks/compare/@airtable/blocks@1.11.0...HEAD)
+
+No changes.
+
+## [1.11.0](https://github.com/airtable/blocks/compare/@airtable/blocks@1.10.2...@airtable/blocks@1.11.0) - 2022-03-10
 
 -   Add workspace id to base object.
 -   Add `Field.updateNameAsync`